Refactor ThriftTests and Java8Tests to use addSourceLines instead of deprecated addSourceFile
Replaces all calls to deprecated CompilationTestHelper#addSourceFile with
addSourceLines, inlining the source code as text blocks. This follows the
same pattern as PR #1426 which handled CoreTests and UnannotatedTests.

Changes:
- ThriftTests: Replace 1 addSourceFile call (Util.java inlined)
- Java8Tests: Replace 4 addSourceFile calls (NullAwayJava8PositiveCases,
  NullAwayJava8NegativeCases, NullAwaySuperFunctionalInterface,
  NullAwayOverrideFunctionalInterfaces all inlined)
- Remove @SuppressWarnings("deprecation") annotations from both classes
- Delete 4 now-unused testdata files for Java8Tests

Fixes #1425 (partial - remaining test files can be addressed in follow-up PRs)
