--- nullaway/src/test/resources/com/uber/nullaway/testdata/NullAwayOverrideFunctionalInterfaces.java
@@ -1,48 +1,0 @@
-/*
- * Copyright (c) 2017 Brandon Mintern
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package com.uber.nullaway.testdata;
-
-import java.util.function.ToIntFunction;
-
-public class NullAwayOverrideFunctionalInterfaces {
-
-  public void test() {
-    call(str -> 42);
-  }
-
-  private int call(ObjToInt<String> f) {
-    return f.call("The answer to life the universe and everything");
-  }
-
-  @FunctionalInterface
-  private static interface ObjToInt<T> extends ObjToIntE<T, RuntimeException>, ToIntFunction<T> {
-    @Override
-    default int applyAsInt(T t) {
-      return call(t);
-    }
-  }
-
-  @FunctionalInterface
-  private static interface ObjToIntE<T, E extends Exception> {
-    int call(T t) throws E;
-  }
-}
